move most of the compiler's kotlin backend into compiler.kotlin
