+++ sqldelight-compiler/src/main/kotlin/app/cash/sqldelight/core/compiler/kotlin/DatabaseExposerGenerator.kt
@@ -1,4 +1,4 @@
-package app.cash.sqldelight.core.compiler
+package app.cash.sqldelight.core.compiler.kotlin
 
 import app.cash.sqldelight.core.SqlDelightFileIndex
 import app.cash.sqldelight.core.lang.ASYNC_RESULT_TYPE
