+++ sqldelight-compiler/src/main/kotlin/app/cash/sqldelight/core/compiler/kotlin/QueriesTypeGenerator.kt
@@ -1,22 +1,24 @@
-package app.cash.sqldelight.core.compiler
+package app.cash.sqldelight.core.compiler.kotlin
 
 import app.cash.sqldelight.core.compiler.model.NamedExecute
 import app.cash.sqldelight.core.compiler.model.NamedMutator
+import app.cash.sqldelight.core.compiler.tryWithElement
+import app.cash.sqldelight.core.lang.*
 import app.cash.sqldelight.core.lang.DRIVER_NAME
 import app.cash.sqldelight.core.lang.DRIVER_TYPE
 import app.cash.sqldelight.core.lang.SUSPENDING_TRANSACTER_IMPL_TYPE
-import app.cash.sqldelight.core.lang.SqlDelightQueriesFile
 import app.cash.sqldelight.core.lang.TRANSACTER_IMPL_TYPE
-import app.cash.sqldelight.core.lang.queriesType
 import app.cash.sqldelight.dialect.api.SqlDelightDialect
 import com.intellij.openapi.module.Module
 import com.squareup.kotlinpoet.FunSpec
 import com.squareup.kotlinpoet.TypeSpec
 
+internal fun SqlDelightQueriesFile.isEmpty() = namedQueries.isEmpty() && namedMutators.isEmpty() && namedExecutes.isEmpty()
+
 class QueriesTypeGenerator(
-  private val module: Module,
-  private val file: SqlDelightQueriesFile,
-  private val dialect: SqlDelightDialect,
+    private val module: Module,
+    private val file: SqlDelightQueriesFile,
+    private val dialect: SqlDelightDialect,
 ) {
   private val generateAsync = file.generateAsync
 
@@ -52,19 +54,19 @@ class QueriesTypeGenerator(
     }
 
     file.namedQueries.forEach { query ->
-      tryWithElement(query.select) {
-        val generator = SelectQueryGenerator(query)
+        tryWithElement(query.select) {
+            val generator = SelectQueryGenerator(query)
 
-        type.addFunction(generator.customResultTypeFunction())
+            type.addFunction(generator.customResultTypeFunction())
 
-        if (query.needsWrapper()) {
-          type.addFunction(generator.defaultResultTypeFunction())
+            if (query.needsWrapper()) {
+                type.addFunction(generator.defaultResultTypeFunction())
+            }
+
+            if (query.needsQuerySubType()) {
+                type.addType(generator.querySubtype())
+            }
         }
-
-        if (query.needsQuerySubType()) {
-          type.addType(generator.querySubtype())
-        }
-      }
     }
 
     file.namedMutators.forEach { mutator ->
@@ -80,16 +82,14 @@ class QueriesTypeGenerator(
   }
 
   private fun TypeSpec.Builder.addExecute(execute: NamedExecute) {
-    tryWithElement(execute.statement) {
-      val generator = if (execute is NamedMutator) {
-        MutatorQueryGenerator(execute)
-      } else {
-        ExecuteQueryGenerator(execute)
-      }
+      tryWithElement(execute.statement) {
+          val generator = if (execute is NamedMutator) {
+              MutatorQueryGenerator(execute)
+          } else {
+              ExecuteQueryGenerator(execute)
+          }
 
-      addFunction(generator.function())
-    }
+          addFunction(generator.function())
+      }
   }
 }
-
-internal fun SqlDelightQueriesFile.isEmpty() = namedQueries.isEmpty() && namedMutators.isEmpty() && namedExecutes.isEmpty()
