+++ sqldelight-compiler/src/test/kotlin/app/cash/sqldelight/core/queries/PgInsertOnConflictTest.kt
@@ -1,6 +1,6 @@
 package app.cash.sqldelight.core.queries
 
-import app.cash.sqldelight.core.compiler.MutatorQueryGenerator
+import app.cash.sqldelight.core.compiler.kotlin.MutatorQueryGenerator
 import app.cash.sqldelight.dialects.postgresql.PostgreSqlDialect
 import app.cash.sqldelight.test.util.FixtureCompiler
 import app.cash.sqldelight.test.util.withUnderscores
