+++ sqldelight-compiler/src/test/kotlin/app/cash/sqldelight/core/queries/async/AsyncSelectQueryTypeTest.kt
@@ -1,8 +1,8 @@
 package app.cash.sqldelight.core.queries.async
 
 import app.cash.sqldelight.core.TestDialect
-import app.cash.sqldelight.core.compiler.ExecuteQueryGenerator
-import app.cash.sqldelight.core.compiler.SelectQueryGenerator
+import app.cash.sqldelight.core.compiler.kotlin.ExecuteQueryGenerator
+import app.cash.sqldelight.core.compiler.kotlin.SelectQueryGenerator
 import app.cash.sqldelight.core.dialects.intType
 import app.cash.sqldelight.core.dialects.textType
 import app.cash.sqldelight.dialects.postgresql.PostgreSqlDialect
